requirements.txt
Flask
flask-cors
requests
pandas
numpy
matplotlib
python-dateutil
scikit-learn
tensorflow
prophet
statsmodels
google-cloud-storage
gunicorn
